added login, register, home and no page in addition to routing to the frontend
