+++ frontend/src/components/Login/Login.jsx
@@ -1,0 +1,7 @@
+import './Login.css'
+
+const Login = () => {
+    return <h1>Login</h1>
+}
+
+export default Login;
