+++ frontend/src/components/NoPage/NoPage.jsx
@@ -1,0 +1,7 @@
+import './NoPage.css'
+
+const NoPage = () => {
+
+}
+
+export default NoPage
