+++ frontend/src/components/Register/Register.jsx
@@ -1,0 +1,7 @@
+import './Register.css'
+
+const Register = () => {
+    return <h1>Register</h1>
+}
+
+export default Register;
